+++ src/Response/Sender.php
@@ -1,8 +1,9 @@
 <?php
 declare(strict_types = 1);
 
-namespace Innmind\Http;
+namespace Innmind\Http\Response;
 
+use Innmind\Http\Response;
 use Innmind\Immutable\{
     Attempt,
     SideEffect,