+++ CHANGELOG.md
@@ -79,6 +79,9 @@
 - `Innmind\Http\Header` constructor is now private, use `::of` named constructor
 - `Innmind\Http\Header\Value` constructor is now private, use `::of` named constructor
 - `Innmind\Http\Header\Parameter` constructor is now private, use `::of` named constructor
+- `Innmind\Http\Sender` as been moved to `Innmind\Http\Response\Sender`
+- `Innmind\Http\ResponseSender` as been moved to `Innmind\Http\Response\Sender\Native`
+- `Innmind\Http\Response\Sender\Native` constructor is now private, use `::of()` named constructor
 
 ### Removed